not print new line for progress in non-term output
we ignore progress data in non-terminal output, so do not print a new line for
those. Also tweak the output in terminal mode because the old logic is wrong
with the server not always returning non-nil Progress in the message.

Docker-DCO-1.1-Signed-off-by: Daniel, Dao Quang Minh <dqminh89@gmail.com> (github: dqminh)