fix race in cli between attach and start container
The race happens when the cli hijacks connection successfully and calls start,
but the server hasnt start the attach job yet ( because it
s still busy deserializing the config). The result is that start finishes
before attach and the cli stops responding

preload cfg to make sure that we can start attach job ASAP because
deserializing large config can be slow.

This makes it less likely that there's a race between attach and jobs that
alters the container's stream such as start.

Docker-DCO-1.1-Signed-off-by: Daniel, Dao Quang Minh <dqminh89@gmail.com> (github: dqminh)