add useCgroup to create exec's subcgroup
when an exec session is created with useCgroup is true, a subcgroup will be
created and all processes belongs to that cgroup will be cleaned up when the
main exec process exits.

When exec session's useCgroup is false, its background processes
will not be killed when the session is stopped. This will pollute the
container's PID namespace, but is useful in some cases where user want to
persist some daemonized processes.

Signed-off-by: Daniel, Dao Quang Minh <dqminh89@gmail.com>